requirements.txt
flask==2.1.0
supabase
python-dotenv
requests
Flask-Limiter
validate-email-address
pyDNS